requirements.txt
requests
websocket-client

Pillow

PyYAML

huggingface_hub
tqdm
